Add tarball dry run

+++ Docker/publish/upload_tar/publish.sh
@@ -24,22 +24,27 @@ echo "Will upload tarballs to $RELEASE_HOST at $RELEASE_FULL_PATH"
 eval "$(ssh-agent -s)"
 ssh-add /root/.ssh/"$RELEASE_SSH_IDENTITY"
 
-# create dir
-ssh "$RELEASE_USER"@"$RELEASE_HOST" mkdir -p "$RELEASE_FULL_PATH"
-
 # move to uploads with release tag
 cp $ARTIFACT_PATH/Wikibase.tar.gz /uploads/wikibase."$RELEASE_VERSION-$WMDE_RELEASE_VERSION".tar.gz
 cp $ARTIFACT_PATH/wdqs-frontend.tar.gz /uploads/wdqs-frontend."$WMDE_RELEASE_VERSION".tar.gz
 
-# upload
-scp /uploads/* "$RELEASE_USER"@"$RELEASE_HOST":"$RELEASE_FULL_PATH"
+if [ -z "$DRY_RUN" ]; then
+    # create dir
+    ssh "$RELEASE_USER"@"$RELEASE_HOST" mkdir -p "$RELEASE_FULL_PATH"
 
-# create dir
-ssh "$RELEASE_USER"@"$RELEASE_HOST" chmod -R g+w "$RELEASE_FULL_PATH"
-ssh "$RELEASE_USER"@"$RELEASE_HOST" chgrp -R releasers-wikibase "$RELEASE_FULL_PATH"
+    # upload
+    scp /uploads/* "$RELEASE_USER"@"$RELEASE_HOST":"$RELEASE_FULL_PATH"
 
-# review dir contents
-ssh "$RELEASE_USER"@"$RELEASE_HOST" ls -ash "$RELEASE_FULL_PATH"
+    # create dir
+    ssh "$RELEASE_USER"@"$RELEASE_HOST" chmod -R g+w "$RELEASE_FULL_PATH"
+    ssh "$RELEASE_USER"@"$RELEASE_HOST" chgrp -R releasers-wikibase "$RELEASE_FULL_PATH"
+
+    # review dir contents
+    ssh "$RELEASE_USER"@"$RELEASE_HOST" ls -ash "$RELEASE_FULL_PATH"
+else
+    echo "DRY RUN! Not uploading anything."
+    ssh "$RELEASE_USER"@"$RELEASE_HOST" ls -ash "$RELEASE_DIR"
+fi
 
 # remove identity
 ssh-add -D
